# Server/api-spec/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.1.0",
        "title": "KFCSmartCorp",
        "description": "HRM System for SmartDEV"
    },
    "host": "localhost:3007",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/role": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Get all role",
                "description": "Gets all list Role objects.\n",
                "responses": {
                    "200": {
                        "description": "List all roles",
                        "schema": {
                            "title": "Object",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/GetRoleApi"
                            }
                        }
                    },
                    "401": {
                        "description": "sends request errors",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Role"
                ],
                "summary": "Create new a role",
                "description": "This api use to create new a role\n",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "ROles object that needs to be added to the store",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RoleApi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user role created",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "sends request errors",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Role"
                ],
                "summary": "Update a role",
                "description": "This api use to update a role\n",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Roles object that needs to be added to the store",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RoleApi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user role created",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "sends request errors",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/role/{id}": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Get role by role id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "type": "string",
                        "description": "ID of the user",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sends the user with user Id",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/GetRoleApi"
                            }
                        }
                    },
                    "401": {
                        "description": "sends request errors",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Role"
                ],
                "summary": "Delete role by role id",
                "description": "This can only be done by the logged in user.",
                "operationId": "deleteUser",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "The userId that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User delete successful",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "sends request errors",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/role/getDataPermissions": {
            "post": {
                "tags": [
                    "Role"
                ],
                "summary": "Creates list of permission with given input array",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "List of user object",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Permission"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "RoleApi": {
            "properties": {
                "username": {
                    "type": "string"
                },
                "type": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Type"
                    }
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Permission"
                    }
                }
            }
        },
        "GetRoleApi": {
            "properties": {
                "_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "type": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Type"
                    }
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Permission"
                    }
                }
            }
        },
        "Type": {
            "properties": {
                "roleTypeId": {
                    "type": "string"
                },
                "nameType": {
                    "type": "string"
                }
            }
        },
        "Permission": {
            "properties": {
                "permission_id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "rights": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Right"
                    }
                }
            }
        },
        "Right": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "right_id": {
                    "type": "string"
                }
            }
        },
...
    }
}